+++ agendamlg-ejb/src/java/app/ejb/EventoFacade.java
@@ -118,7 +118,7 @@ public class EventoFacade extends AbstractFacade<Evento> {
     }
 
     
-    public List<Evento> buscarEventoCategorias(List<Categoria> categorias, Usuario usuario, boolean filtroCercania, int x, int y, int radio){
+    public List<Evento> buscarEventoCategorias(List<Categoria> categorias, Usuario usuario, boolean filtroCercania, double x, double y, double radio){
 
         /* Query q = this.em.createQuery("select distinct e from Evento e join e.categoriaList c where c in :categorias");
         q.setParameter("categorias", categorias);
@@ -202,15 +202,15 @@ public class EventoFacade extends AbstractFacade<Evento> {
     // Métodos auxiliares
     
     // Dada una ubicacion (x,y) y un evento devuelve la distancia hasta ese evento
-    private double distanciaAEvento(int x, int y, Evento evento) {
+    private double distanciaAEvento(double x, double y, Evento evento) {
         // Obtener coordenadas x e y del evento
         String[] coordenadas = evento.getDireccion().split(",");
-        int eventoX = Integer.parseInt(coordenadas[0]);
-        int eventoY = Integer.parseInt(coordenadas[1]);
+        double eventoX = Double.parseDouble(coordenadas[0]);
+        double eventoY = Double.parseDouble(coordenadas[1]);
         return distanciaEntreDosPuntos(x, y, eventoX, eventoY);
     }
     // Dados dos punto de la forma (x1,y1),(x2,y2) calcula la distancia entre ambos
-    private double distanciaEntreDosPuntos(int x1, int y1, int x2, int y2) {
+    private double distanciaEntreDosPuntos(double x1, double y1, double x2, double y2) {
         return Math.sqrt(Math.pow((x1 - x2), 2) + Math.pow((y1 - y2), 2));
     }
